requirements.txt

# Тестирование API
requests==2.33.1
pytest==9.0.2
pytest-html==4.1.1
pytest-cov==4.1.0
responses==0.26.0

# Тестирование UI (веб)
selenium==4.40.0
webdriver-manager==4.0.2
playwright==1.40.0

# Дополнительные утилиты
python-dotenv==1.0.0
loguru==0.7.2
allure-pytest==2.13.2